requirements.txt
Adafruit-BBIO==0.0.19
Django==1.5.2
RPi.GPIO==0.5.4
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
South==0.8.2
Sphinx==1.1.3
Unipath==1.0
argparse==1.2.1
boto==2.11.0
coverage==3.6
dj-database-url==0.2.2
dj-static==0.0.5
django-debug-toolbar==0.9.4
django-discover-runner==1.0
django-extensions==1.2.0
django-gevent-deploy==1.0
django-nose==1.2
django-storages==1.1.8
django-toolbelt==0.0.1
docutils==0.11
gevent==0.13.8
greenlet==0.4.1
gunicorn==18.0
nose==1.3.0
psycopg2==2.5.1
selenium==2.35.0
six==1.3.0
static==0.4
twilio==3.6.6
wsgiref==0.1.2
pytz==2013b

